+++ scripts/bert/run_pretraining_hvd.py
@@ -49,7 +49,7 @@ from gluonnlp.data import SimpleDatasetStream, FixedBucketSampler, NumpyDataset,
 from utils import profile
 from fp16_utils import FP16Trainer
 from pretraining_utils import get_model, get_pretrain_dataset, get_dummy_dataloader
-from pretraining_utils import save_params, log
+from pretraining_utils import save_params, log, split_and_load, evaluate
 
 parser = argparse.ArgumentParser(description='BERT pretraining example.')
 parser.add_argument('--num_steps', type=int, default=20, help='Number of optimization steps')
@@ -114,16 +114,6 @@ logging.debug('local_rank = %d, rank = %d, num_workers = %d',
               local_rank, rank, num_workers)
 logging.info(nlp)
 
-def split_and_load(arrs, ctx):
-    """split and load arrays to a list of contexts"""
-    assert isinstance(arrs, (list, tuple))
-    if len(ctx) == 1:
-        return [[arr.as_in_context(ctx[0]) for arr in arrs]]
-    else:
-        # split and load
-        loaded_arrs = [gluon.utils.split_and_load(arr, ctx, even_split=False) for arr in arrs]
-        return zip(*loaded_arrs)
-
 def forward(data, model, mlm_loss, nsp_loss, vocab_size):
     """forward computation for evaluation"""
     (input_id, masked_id, masked_position, masked_weight, \
@@ -173,64 +163,6 @@ class ParallelBERT(Parallelizable):
             ls.backward()
         return ls, next_sentence_label, classified, masked_id, decoded, \
                masked_weight, ls1, ls2, valid_length
-
-def evaluate(data_eval, model, nsp_loss, mlm_loss, vocab_size, ctx):
-    """Evaluation function."""
-    mlm_metric = MaskedAccuracy()
-    nsp_metric = MaskedAccuracy()
-    mlm_metric.reset()
-    nsp_metric.reset()
-
-    eval_begin_time = time.time()
-    begin_time = time.time()
-    step_num = 0
-    local_mlm_loss = local_nsp_loss = 0
-    total_mlm_loss = total_nsp_loss = 0
-    local_num_tks = 0
-    for _, dataloader in enumerate(data_eval):
-        for _, data in enumerate(dataloader):
-            step_num += 1
-
-            data_list = split_and_load(data, ctx)
-            loss_list = []
-            ns_label_list, ns_pred_list = [], []
-            mask_label_list, mask_pred_list, mask_weight_list = [], [], []
-            for data in data_list:
-                out = forward(data, model, mlm_loss, nsp_loss, vocab_size)
-                (ls, next_sentence_label, classified, masked_id,
-                 decoded, masked_weight, ls1, ls2, valid_length) = out
-                loss_list.append(ls)
-                ns_label_list.append(next_sentence_label)
-                ns_pred_list.append(classified)
-                mask_label_list.append(masked_id)
-                mask_pred_list.append(decoded)
-                mask_weight_list.append(masked_weight)
-
-                local_mlm_loss += ls1.as_in_context(mx.cpu())
-                local_nsp_loss += ls2.as_in_context(mx.cpu())
-                local_num_tks += valid_length.sum().as_in_context(mx.cpu())
-            nsp_metric.update(ns_label_list, ns_pred_list)
-            mlm_metric.update(mask_label_list, mask_pred_list, mask_weight_list)
-
-            # logging
-            if (step_num + 1) % (args.log_interval) == 0:
-                total_mlm_loss += local_mlm_loss
-                total_nsp_loss += local_nsp_loss
-                log(begin_time, local_num_tks, local_mlm_loss, local_nsp_loss,
-                    step_num, mlm_metric, nsp_metric, None, args.log_interval)
-                begin_time = time.time()
-                local_mlm_loss = local_nsp_loss = local_num_tks = 0
-                mlm_metric.reset_local()
-                nsp_metric.reset_local()
-
-    mx.nd.waitall()
-    eval_end_time = time.time()
-    total_mlm_loss /= step_num
-    total_nsp_loss /= step_num
-    logging.info('mlm_loss={:.3f}\tmlm_acc={:.1f}\tnsp_loss={:.3f}\tnsp_acc={:.1f}\t'
-                 .format(total_mlm_loss.asscalar(), mlm_metric.get_global()[1] * 100,
-                         total_nsp_loss.asscalar(), nsp_metric.get_global()[1] * 100))
-    logging.info('Eval cost={:.1f}s'.format(eval_end_time - eval_begin_time))
 
 def train(data_train, model, nsp_loss, mlm_loss, vocab_size, ctx):
     """Training function."""
@@ -389,12 +321,15 @@ if __name__ == '__main__':
             os.makedirs(ckpt_dir)
 
     if args.data:
+        num_parts = 1 if args.dummy_data_len else num_workers
+        part_idx = 0 if args.dummy_data_len else rank
         data_train = get_pretrain_dataset(args.data, args.batch_size, len(ctx), True,
                                           args.use_avg_len, args.num_buckets,
-                                          num_parts=1 if args.dummy_data_len else num_workers,
-                                          part_idx=0 if args.dummy_data_len else rank,
+                                          num_parts=num_parts, part_idx=part_idx,
                                           prefetch=not args.dummy_data_len)
         train(data_train, model, nsp_loss, mlm_loss, len(tokenizer.vocab), ctx)
     if args.data_eval:
-        data_eval = get_dataset(args.data_eval, args.batch_size_eval, len(ctx), False, False, 1)
-        evaluate(data_eval, model, nsp_loss, mlm_loss, len(tokenizer.vocab), ctx)
+        data_eval = get_pretrain_dataset(args.data_eval, args.batch_size_eval, len(ctx),
+                                         False, False, 1)
+        evaluate(data_eval, model, nsp_loss, mlm_loss, len(tokenizer.vocab), ctx,
+                 args.log_interval, args.dtype)
